MAINT Re-anchor the certificate to the tile-local period replay and measure ULP in the working dtype
The candidate certificate seals `src/_lcm/solution/period_replay.py`, which changed when ride-along NB-EGM periods moved into one tile-local core. The generated source inventory and the independent seal table are regenerated against the current source; the four seals re-anchored for the per-period gated-edge references (`processing.py`, `engine.py`, `simulation/compile.py`, `simulation/simulate.py`) arrive through the merge of `fix/gated-edge-reference-periods`.

The ULP helper measured spacing in the dtype of `operand_magnitude`, so a magnitude promoted to float64 reported float64 spacing for float32 arrays and a one-spacing float32 gap read as 2^28 ULP. Spacing is measured in the compared arrays' own dtype. The interval-batch test replaces a fixed 1e-12 tolerance, which lies below float32 resolution, with `assert_agrees_to_ulp` and the operand magnitude, the partition instrument the batch-size invariance is stated in.

+++ tests/conftest.py
@@ -215,7 +215,11 @@ def assert_agrees_to_ulp(
     gap = np.where(finite, np.abs(got_arr - expected_arr), 0.0)
     magnitude = np.maximum(np.abs(got_arr), np.abs(expected_arr))
     if operand_magnitude is not None:
-        magnitude = np.maximum(magnitude, np.asarray(operand_magnitude))
+        # Measure the spacing in the compared arrays' own format: promoting the
+        # magnitude to a wider float would report the wider format's spacing.
+        magnitude = np.maximum(
+            magnitude, np.asarray(operand_magnitude, dtype=magnitude.dtype)
+        )
     spacing = np.spacing(magnitude)
     in_ulp = np.divide(
         gap, spacing, out=np.zeros(gap.shape, dtype=float), where=gap > 0.0
